+++ vim/pack/plugins/start/fzf/CHANGELOG.md
@@ -1,6 +1,21 @@
 CHANGELOG
 =========
 
+0.43.0
+------
+- `--listen` server can be secured by setting `$FZF_API_KEY` environment
+  variable.
+  ```sh
+  export FZF_API_KEY="$(head -c 32 /dev/urandom | base64)"
+
+  # Server
+  fzf --listen 6266
+
+  # Client
+  curl localhost:6266 -H "x-api-key: $FZF_API_KEY" -d 'change-query(yo)'
+  ```
+- Added `toggle-header` action
+
 0.42.0
 ------
 - Added new info style: `--info=right`